migrate tool and demo to latest v7
