+++ demo/package.json
@@ -7,7 +7,7 @@
     "@ledgerhq/hw-transport-u2f": "^4.63.2",
     "@ledgerhq/hw-transport-web-ble": "^4.63.2",
     "@ledgerhq/hw-transport-webusb": "^4.63.2",
-    "@ledgerhq/live-common": "^7.0.0-beta.0",
+    "@ledgerhq/live-common": "^7.0.0-beta.10",
     "axios": "^0.19.0",
     "babel-polyfill": "^6.26.0",
     "bignumber.js": "^9.0.0",
