--- demo/src/demos/countervalue-intermediary/components/App.js
@@ -1,4 +1,0 @@
-// @flow
-import CounterValuesAPIDemo from "./CounterValuesAPIDemo";
-
-export default CounterValuesAPIDemo;
