--- demo/src/demos/countervalue-intermediary/store.js
@@ -1,15 +1,0 @@
-// @flow
-import { createStore, applyMiddleware, combineReducers } from "redux";
-import { composeWithDevTools } from "redux-devtools-extension";
-import thunkMiddleware from "redux-thunk";
-import app from "./reducers/app";
-import CounterValues from "./countervalues";
-
-export const initStore = () =>
-  createStore(
-    combineReducers({
-      app,
-      countervalues: CounterValues.reducer
-    }),
-    composeWithDevTools(applyMiddleware(thunkMiddleware))
-  );
