+++ demo/src/demos/countervalues/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+import App from "./components/App";
+
+const Demo = () => <App />;
+
+Demo.demo = {
+  title: "Countervalues",
+  url: "/countervalues"
+};
+
+export default Demo;
