+++ demo/src/demos/index.js
@@ -1,5 +1,4 @@
-import CountervalueDirect from "./countervalue-direct";
-import CountervalueIntermediary from "./countervalue-intermediary";
+import Countervalues from "./countervalues";
 import LogsViewer from "./logsviewer";
 import Derivations from "./derivations";
 import Currencies from "./Currencies";
@@ -12,8 +11,7 @@ import Partners from "./Partners";
 import erc20 from "./erc20";
 
 export default {
-  CountervalueDirect,
-  CountervalueIntermediary,
+  Countervalues,
   LogsViewer,
   Derivations,
   Currencies,
