+++ demo/src/index.js
@@ -2,11 +2,14 @@ import "babel-polyfill";
 import "./live-common-setup";
 import React, { Component } from "react";
 import ReactDOM from "react-dom";
+import { Provider } from "react-redux";
 import { BrowserRouter, Route, Link, Switch } from "react-router-dom";
+import { getCountervalues } from "@ledgerhq/live-common/lib/countervalues";
 import { version } from "@ledgerhq/live-common/package.json";
 import "./index.css";
 // import registerServiceWorker from "./registerServiceWorker";
 import Demos from "./demos";
+import { initStore } from "./store";
 
 class Dashboard extends Component {
   render() {
@@ -35,17 +38,49 @@ class Dashboard extends Component {
   }
 }
 
-class App extends Component {
+class App extends Component<*, *> {
+  constructor() {
+    super();
+    const CounterValues = getCountervalues();
+    const store = initStore();
+
+    // quick way to store countervalues with localStorage
+    const LS_KEY = "countervalues_intermediary";
+    try {
+      const json = localStorage.getItem(LS_KEY);
+      if (json) {
+        store.dispatch(CounterValues.importAction(JSON.parse(json)));
+      }
+    } catch (e) {
+      console.warn(e);
+    }
+    store.subscribe(() => {
+      localStorage.setItem(
+        LS_KEY,
+        JSON.stringify(CounterValues.exportSelector(store.getState()))
+      );
+    });
+
+    this.state = {
+      CounterValues,
+      store
+    };
+  }
   render() {
+    const { CounterValues, store } = this.state;
     return (
-      <Switch>
-        <Route exact path="/" component={Dashboard} />
-        {Object.keys(Demos).map(key => {
-          const Demo = Demos[key];
-          const { url } = Demo.demo;
-          return <Route key={key} path={url} component={Demo} />;
-        })}
-      </Switch>
+      <Provider store={store}>
+        <CounterValues.PollingProvider>
+          <Switch>
+            <Route exact path="/" component={Dashboard} />
+            {Object.keys(Demos).map(key => {
+              const Demo = Demos[key];
+              const { url } = Demo.demo;
+              return <Route key={key} path={url} component={Demo} />;
+            })}
+          </Switch>
+        </CounterValues.PollingProvider>
+      </Provider>
     );
   }
 }
