+++ demo/src/store.js
@@ -1,15 +1,15 @@
 // @flow
 import { createStore, applyMiddleware, combineReducers } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
+import { getCountervalues } from "@ledgerhq/live-common/lib/countervalues";
 import thunkMiddleware from "redux-thunk";
 import markets from "./reducers/markets";
-import CounterValues from "./countervalues";
 
 export const initStore = () =>
   createStore(
     combineReducers({
       markets,
-      countervalues: CounterValues.reducer
+      countervalues: getCountervalues().reducer
     }),
     composeWithDevTools(applyMiddleware(thunkMiddleware))
   );
