+++ tool/package.json
@@ -12,7 +12,7 @@
     "@ledgerhq/hw-transport-mocker": "^4.63.2",
     "@ledgerhq/hw-transport-node-ble": "^4.63.2",
     "@ledgerhq/hw-transport-node-hid": "^4.63.2",
-    "@ledgerhq/ledger-core": "^3.0.0-beta.1",
+    "@ledgerhq/ledger-core": "^3.0.0-beta.6",
     "@ledgerhq/live-common": "^7.0.0-beta.10",
     "@ledgerhq/logs": "^4.62.0",
     "axios": "^0.19.0",
